renamed build to model and required it from serv
