--- build
@@ -1,2 +1,0 @@
-#!/usr/bin/env bash
-node build.js
