+++ serv.js
@@ -23,94 +23,10 @@ var http = require("http"),
         path = require("path"),
         fs = require("fs"),
         walk = require('fs-walk'),
-        model = require('./build'),
+        model = require('./model'),
         port = process.argv[2] || 3000,
         defaultThemesDir = '/node_modules/reveal.js/css/theme/',
         templatesDir = '/templates/';
-
-
-/* build model.json */
-
-var model = {};
-
-function defaultThemes(items, dir) {
-    return items.filter(function (el) {
-        return el.endsWith('.css');
-    }).map(function (el) {
-        return {
-            title: el.replace(".css", ""),
-            file: dir + el
-        };
-    });
-}
-
-//themes. first personal themes then default themes
-function themes() {
-    model.themes = defaultThemes(fs.readdirSync("themes"), "/themes/");
-    model.theme = model.themes[0].file; //TODO failes when nothing there
-    defaultThemes(fs.readdirSync("." + defaultThemesDir), defaultThemesDir).forEach(function (entry) {
-        model.themes.push(entry);
-    });
-}
-
-function processTemplates(items) {
-    return items.filter(function (el) {
-        return el.endsWith('.html');
-    }).map(function (el) {
-        return {
-            title: el.replace(".html", ""),
-            file: path.join("/", templatesDir, el)
-        };
-    });
-}
-
-function templates() {
-//templates
-    model.templates = processTemplates(fs.readdirSync("templates"));
-
-    model.templates.forEach(function (data) {
-        if (data.title === "ivonet") {
-            model.template = data.file;
-        }
-    });
-    if (model.template === undefined) {
-        model.template = model.templates[0].file;
-    }
-}
-function transitions() {
-//Transitions
-    model.transitions = [
-        "none",
-        "fade",
-        "slide",
-        "convex",
-        "concave",
-        "zoom"
-    ];
-    model.transition = 'convex';
-}
-function presentations() {
-//Presentations
-    model.slides = [];
-    walk.walkSync('./slides', function (basedir, filename, stat) {
-        "use strict";
-        if (filename.endsWith(".md")) {
-            var slide = {};
-            slide.title = filename.replace(".md", "");
-            slide.file = path.join("/", basedir, filename);
-            model.slides.push(slide)
-        }
-    });
-}
-function buildModel() {
-    themes();
-    templates();
-    transitions();
-    presentations();
-    console.log(model);
-    return JSON.stringify(model);
-}
-
 
 function endsWith(str, suffix) {
     return str.indexOf(suffix, str.length - suffix.length) !== -1;
@@ -127,6 +43,16 @@ http.createServer(function (request, response) {
         '.js': "text/javascript",
         '.json': "application/json"
     };
+    if (endsWith(filename, 'model.json')) {
+        var headers = {};
+        var contentType = contentTypesByExtension[path.extname(filename)];
+        if (contentType)
+            headers["Content-Type"] = contentType;
+        response.writeHead(200, headers);
+        response.write(model.buildModel(), "binary");
+        response.end();
+        return;
+    }
     fs.exists(filename, function (exists) {
         if (!exists) {
             response.writeHead(404, {"Content-Type": "text/plain"});
@@ -135,16 +61,6 @@ http.createServer(function (request, response) {
             return;
         }
 
-        if (endsWith(filename, 'model.json')) {
-            var headers = {};
-            var contentType = contentTypesByExtension[path.extname(filename)];
-            if (contentType)
-                headers["Content-Type"] = contentType;
-            response.writeHead(200, headers);
-            response.write(model.buildModel(), "binary");
-            response.end();
-            return;
-        }
         if (fs.statSync(filename).isDirectory())
             filename += '/index.html';
